requirements.txt
autoflake==2.1.1    
autopep8==2.0.2    
black==23.3.0   
cryptography==40.0.2
celery==5.2.7
dj-rest-auth==3.0.0    
Django==4.2
django_celery_beat==2.5.0      
django-allauth==0.54.0   
django-bmemcached==0.3.0    
django-cors-headers==3.14.0   
django-model-utils==4.3.1    
django-notifications-hq==1.7.0    
djangorestframework==3.14.0   
djangorestframework-simplejwt==5.2.2
isort==5.12.0     
Pillow==9.4.0  
psycopg2    
pyflakes==3.0.1    
PyJWT==2.6.0    
pymemcache==4.0.0    
python-binary-memcached==0.31.2   
python-dateutil==2.8.2    
python-dotenv==1.0.0
pytest==7.3.1  
pytest_django==4.5.2
pycryptodome==3.17
requests==2.28.2 
redis==4.5.4
six==1.16.0